requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
datasets==3.2.0
dill==0.3.8
distro==1.9.0
einops==0.8.0
einops-exts==0.0.4
filelock==3.13.1
flamingo-pytorch==0.1.2
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2024.6.1
h11==0.14.0
h5py==3.12.1
histogpt @ git+https://github.com/marrlab/HistoGPT@98ecaecf63d39b89fd2c391bffdbfc5543322995
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
numpy==2.1.2
openai==1.61.0
opencv-python==4.11.0.86
openslide-python==1.4.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
psutil==6.1.1
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-triton-rocm==3.2.0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
sacremoses==0.1.1
safetensors==0.5.2
setuptools==70.2.0
six==1.17.0
slideio==2.6.5
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0+rocm6.2.4
torchaudio==2.6.0+rocm6.2.4
torchvision==0.21.0+rocm6.2.4
tqdm==4.67.1
transformers==4.48.2
trl==0.14.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
